Refactored metrics code
