+++ wazimap_ng/profile/serializers/metrics_serializer.py
@@ -4,64 +4,24 @@ from wazimap_ng.datasets.models import IndicatorData
 
 from .. import models
 
-def get_subindicator(metric):
-    subindicators = metric.indicator.subindicators
-    idx = metric.subindicator if metric.subindicator is not None else 0
-    return subindicators[idx]
+from .helpers import get_subindicator, get_sum, MetricCalculator
 
 def get_indicator_data(profile_key_metric, geographies):
     indicator_data = IndicatorData.objects.filter(indicator__profilekeymetrics=profile_key_metric, geography__in=geographies)
     return indicator_data
 
-def get_sum(data, group=None, subindicator=None):
-    if (group is not None and subindicator is not None):
-        return sum([float(row["count"]) for row in data if group in row and row[group] == subindicator])
-    return sum(float(row["count"]) for row in data)
-
 def sibling(profile_key_metric, geography):
     siblings = geography.get_siblings()
     data = get_indicator_data(profile_key_metric, siblings)
-    group = profile_key_metric.indicator.groups[0]
-    subindicator = get_subindicator(profile_key_metric)
-    numerator = None
-    denominator = 0
-    total = 0
-    for datum in data:
-        total += get_sum(datum.data)
-        if datum.geography == geography:
-            geography_total = get_sum(datum.data)
-
-    denominator, numerator = total, geography_total        
-
-    if denominator > 0 and numerator is not None:
-        return numerator / denominator
+    return MetricCalculator.sibling(data, profile_key_metric, geography)
 
 def absolute_value(profile_key_metric, geography):
     data = get_indicator_data(profile_key_metric, [geography]).first().data
-    group = profile_key_metric.indicator.groups[0]
-    
-    subindicator = get_subindicator(profile_key_metric)
-    filtered_data = [row["count"] for row in data if group in row and row[group] == subindicator]
-        
-    return get_sum(data, group, subindicator)
+    return MetricCalculator.absolute_value(data, profile_key_metric, geography)
 
 def subindicator(profile_key_metric, geography):
     data = get_indicator_data(profile_key_metric, [geography]).first().data
-    group = profile_key_metric.indicator.groups[0]
-
-    subindicator = get_subindicator(profile_key_metric)
-    numerator = get_sum(data, group, subindicator)
-    denominator = get_sum(data)
-    
-
-    if denominator > 0 and numerator is not None:
-        return numerator / denominator
-
-algorithms = {
-    "absolute_value": absolute_value,
-    "sibling": sibling,
-    "subindicators": subindicator
-}
+    return MetricCalculator.subindicator(data, profile_key_metric, geography)
 
 def MetricsSerializer(profile, geography):
     out_js = {}
